requirements/base.txt
pytz==2018.7  # https://github.com/stub42/pytz
python-slugify==1.2.6  # https://github.com/un33k/python-slugify
Pillow==5.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.3.0  # https://github.com/hynek/argon2_cffi
redis>=2.10.5  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==2.0.9  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.38.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.0  # https://github.com/encode/django-rest-framework
django-rest-auth==0.9.3
coreapi==2.3.3  # https://github.com/core-api/python-client

# tratum
django-embed-video==1.1.2
django-import-export==1.0.1
django-mptt==0.9.0
django-mptt-urls==2.0.3
git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
pdfkit==0.6.1
num2words==0.5.7
raven==6.9.0  # https://github.com/getsentry/raven-python

astroid==1.6.4
cairocffi==0.8.1
CairoSVG==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
cssselect2==0.2.1
mccabe==0.6.1
et-xmlfile==1.0.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
lxml==4.2.1
num2words==0.5.7
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.4
pdfkit==0.6.1
pdfrw==0.4
sqlparse==0.2.4
reportlab==3.4.0
requests==2.19.1
requests-oauthlib==1.0.0
tablib==0.12.1
pypandoc==1.4
PyPDF2==1.26.0
Pyphen==0.9.4
python-docx==0.8.6
python3-openid==3.1.0
PyYAML==3.13
xhtml2pdf==0.2.2
xlrd==1.1.0
xlwt==1.3.0



